+++ lib/puppet-check/cli.rb
@@ -21,7 +21,7 @@ class PuppetCheck::CLI
 
       # base options
       opts.on('--version', 'Display the current version.') do
-        puts 'puppet-check 1.2.0'
+        puts 'puppet-check 1.2.1'
         exit 0
       end